🔧 configure metadata for SSG
