+++ webapp/src/app/health/page.tsx
@@ -1,18 +1,9 @@
-"use client";
+import Health from "@/app/health/health";
 
-import { useQuery } from "@apollo/client";
-import { gql } from "@/gql/__generated__";
+export const metadata = {
+  title: "Health - Example Page",
+};
 
-export default function Health() {
-  const Query = gql(`
-  query HealthCheckQuery { healthCheck }
-  `);
-
-  // call for graphql query
-  const { data } = useQuery(Query);
-  return (
-    <div>
-      <h1>Health: {data?.healthCheck}</h1>
-    </div>
-  );
+export default function Page() {
+  return <Health />;
 }
