+++ webapp/src/app/order/allergy/allergy.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import React from "react";
 
 function Page() {
