+++ webapp/src/app/order/allergy/page.tsx
@@ -1,0 +1,9 @@
+import Allergy from "@/app/order/allergy/allergy";
+
+export const metadata = {
+  title: "Title - CHANGE ME!!!",
+};
+
+export default function Page() {
+  return <Allergy />;
+}
