+++ webapp/src/app/order/menu/[id]/page.tsx
@@ -2,6 +2,10 @@ export function generateStaticParams() {
   return [{ id: "1" }, { id: "2" }, { id: "3" }];
 }
 
+export const metadata = {
+  title: "Title - CHANGE ME!!!",
+};
+
 export default async function Page({
   params,
 }: {
